requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
dnspython==2.1.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
pymongo==3.11.3
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1
